requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
typer[all]>=0.9.0
rich>=13.7.0
flask>=3.0.0
